requirements.txt
Flask
nbtlib
mcstatus
python-dotenv
paramiko